add selectize macros
